enqueue message with 4 goroutine

+++ structs.go
@@ -7,7 +7,7 @@ import "time"
 //DriverLocation contain the position of one driver
 type DriverLocation struct {
 	DriverID    int       `json:"driverID"`
-	Latitude    float32   `json:"latitude"`
-	Longitude   float32   `json:"longitude"`
+	Latitude    float64   `json:"latitude"`
+	Longitude   float64   `json:"longitude"`
 	CreatedDate time.Time `json:"createdDate"`
 }
